requirements.txt
streamlit>=1.28.0
google-cloud-bigquery>=3.11.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
slack-sdk>=3.20.0
python-dotenv>=0.19.0
